improved duplicate code in hover, autocomplete, and pyodide
