+++ src/utils/GetAttUtils.ts
@@ -1,5 +1,9 @@
 import { Position } from 'vscode-languageserver-protocol';
 import { Context } from '../context/Context';
+import { SchemaRetriever } from '../schema/SchemaRetriever';
+import { LoggerFactory } from '../telemetry/LoggerFactory';
+
+const log = LoggerFactory.getLogger('GetAttUtils');
 
 /**
  * Determines the position in GetAtt arguments (1 for resource, 2 for attribute)
@@ -97,4 +101,39 @@ export function extractAttributeName(args: unknown, context: Context): string | 
     }
 
     return context.text;
+}
+
+/**
+ * Gets documentation for a resource attribute from the schema.
+ * Returns the attribute description if found in the schema, otherwise returns a default description.
+ */
+export function getAttributeDocumentationFromSchema(
+    schemaRetriever: SchemaRetriever,
+    resourceType: string,
+    attributeName: string,
+): string {
+    const schema = schemaRetriever.getDefault().schemas.get(resourceType);
+
+    // Provide fallback description even when schema is not available
+    let description = `**${attributeName}** attribute of **${resourceType}**\n\nReturns the value of this attribute when used with the GetAtt intrinsic function.`;
+
+    if (schema) {
+        const jsonPointerPath = `/properties/${attributeName.replaceAll('.', '/')}`;
+
+        try {
+            const resolvedSchemas = schema.resolveJsonPointerPath(jsonPointerPath);
+
+            if (resolvedSchemas.length > 0) {
+                const firstSchema = resolvedSchemas[0];
+
+                if (firstSchema.description) {
+                    description = firstSchema.description;
+                }
+            }
+        } catch (error) {
+            log.debug({ error, resourceType, attributeName }, 'Error resolving attribute documentation');
+        }
+    }
+
+    return description;
 }
